+++ public_html/templates/footer.php
@@ -25,6 +25,7 @@
 
     <script src="/js/materialize.min.js?v=<?= urlencode($assetVersion) ?>"></script>
     <script src="/assets/whois.js?v=<?= urlencode($assetVersion) ?>"></script>
+    <script src="/assets/vt.js?v=<?= urlencode($assetVersion) ?>"></script>
     <script src="/js/app.js?v=<?= urlencode($assetVersion) ?>"></script>
 </body>
 </html>